Added the changes requested by reviewer,updated the unit tests,removed duplicate files inside the Xcode folder.
